Extract New York Craigslist Jobs
Note: this script was performed on jupyter notebook and used with the connection string. However, for the purpose of security, it is shown to use config.json

Due to the homework requirements needing to have more than 7500 rows of data, the web scraping was redone to include job postings in the New York area. No longer limited to Long Island because as seen in the previous script of Extract_Craigslist, there were only 349 records in the last month. Now there have been 19464 job postings extracted in the last month, from April 13th to May 14th